dbeaver/pro#1823 move to DesktopApplicationImpl
